requirements.txt
azure-ai-documentintelligence==1.0.2
azure-search-documents==11.6.0
azure-storage-blob==12.19.0
openai==1.54.0
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
